requirements.txt
discord.py==2.3.2
discord-webhook==1.3.0
aiohttp==3.9.1
openai>=1.30.0
flask==3.0.0
